fix(message): reject protected channelData fragments
